backend/requirements.txt
flask==3.0.0
flask-cors==5.0.1
python-dotenv==1.0.1
requests==2.32.3
langchain==0.3.3
langchain-community==0.3.2
langchain-google-genai==2.0.1
google-generativeai==0.8.4
langchain-google-vertexai==2.0.5
google-cloud-aiplatform==1.70.0
chromadb==0.5.13
pydantic==2.9.2
werkzeug==3.0.4
gunicorn==23.0.0
numpy==1.26.4
pandas==2.2.3
python-docx==1.1.2
pypdf==5.1.0
PyJWT==2.9.0
mysql-connector-python==9.1.0
nltk==3.9.1